Move ComponentInfo and DetectorInfo PythonIterators to geometry
Re #26334
