+++ Framework/PythonInterface/mantid/geometry/inc/MantidPythonInterface/geometry/ComponentInfoPythonIterator.h
