+++ Framework/PythonInterface/mantid/geometry/inc/MantidPythonInterface/geometry/DetectorInfoPythonIterator.h
