+++ Framework/PythonInterface/mantid/geometry/src/Exports/ComponentInfoPythonIterator.cpp
@@ -4,7 +4,7 @@
 //     NScD Oak Ridge National Laboratory, European Spallation Source
 //     & Institut Laue - Langevin
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "MantidPythonInterface/api/ComponentInfoPythonIterator.h"
+#include "MantidPythonInterface/geometry/ComponentInfoPythonIterator.h"
 #include "MantidPythonInterface/core/VersionCompat.h"
 #include <boost/python/class.hpp>
 #include <boost/python/iterator.hpp>
